+++ cli/internal/api/api_test.go
@@ -134,7 +134,8 @@ func TestStreamLogs(t *testing.T) {
 		name                  string
 		apiRespBody           []byte
 		apiRespStatusCode     int
-		endpoint              string          // Used to create new request error.
+		endpoint              string // Used to create new request error.
+		loggedBytes           int64
 		mockHTTPClient        *mockHTTPClient // Only used when needed.
 		writeBadContentLength bool            // Used to create response body error.
 		want                  []byte
@@ -145,6 +146,13 @@ func TestStreamLogs(t *testing.T) {
 			apiRespBody:       readFile(t, "stream_logs_good.txt"),
 			apiRespStatusCode: http.StatusOK,
 			want:              readFile(t, "stream_logs_good.txt"),
+		},
+		{
+			name:              "Log after 8 bytes",
+			apiRespBody:       readFile(t, "stream_logs_good.txt"),
+			apiRespStatusCode: http.StatusOK,
+			loggedBytes:       8,
+			want:              readFile(t, "stream_logs_good.txt")[8:],
 		},
 		{
 			name:              "error non-200 response",
@@ -207,7 +215,7 @@ func TestStreamLogs(t *testing.T) {
 			}
 
 			var b bytes.Buffer
-			err := client.StreamLogs(context.Background(), &b, "workflow1")
+			err := client.StreamLogs(context.Background(), &b, "workflow1", tt.loggedBytes)
 
 			if tt.wantErr != nil {
 				assert.EqualError(t, err, tt.wantErr.Error())